Fix deleting temp video when not using experimental picker

+++ lib/pages/save_video/save_video_page.dart
@@ -302,8 +302,11 @@ class _SaveVideoPageState extends State<SaveVideoPage> {
   }
 
   Future<void> closePopupAndPushToRecording(String cacheVideoPath) async {
-    // Deleting video from cache
-    StorageUtils.deleteFile(cacheVideoPath);
+    final isExperimentalPicker = SharedPrefsUtil.getBool('useExperimentalPicker') ?? true;
+    if (!isExperimentalPicker) {
+      // Deleting video from cache
+      StorageUtils.deleteFile(cacheVideoPath);
+    }
     _trimmer.videoPlayerController?.dispose();
     Get.back();
     final sdkVersion = SharedPrefsUtil.getInt('sdkVersion');
@@ -441,9 +444,11 @@ class _SaveVideoPageState extends State<SaveVideoPage> {
         // Prevent showing the option to re-record video if not coming from the recording page
         final isFromRecordingPage = routeArguments['isFromRecordingPage'];
         if (!isFromRecordingPage) {
-          // Deleting video from cache
-          // TODO: delete if wechat picker is kept
-          // StorageUtils.deleteFile(_tempVideoPath);
+          final isExperimentalPicker = SharedPrefsUtil.getBool('useExperimentalPicker') ?? true;
+          if (!isExperimentalPicker) {
+            // Deleting video from cache
+            StorageUtils.deleteFile(_tempVideoPath);
+          }
           Get.back();
         } else {
           showDialog(
